Create README.md
SOLIDITY SUBMISSION
In simpler terms, this submission is a proposal for a Solidity project aimed at MetaCrafter. The project involves creating a smart contract that can increase the value associated with an Ethereum address and also stop the process if the amount being added exceeds the address's balance.

DESCRIPTION
In simpler terms, this program is coded in Solidity, a programming language for creating smart contracts on Ethereum, a prominent blockchain technology. This contract is a straightforward piece of code that adds ether to your account and tracks it. If the deduction amount exceeds what's available, it won't deduct. Additionally, it displays the current balance in your address.

RUN THE CODE
// SPDX-License-Identifier: MIT
pragma solidity 0.8.18;

contract MyToken {

    // public variables here
    string public tokenName="META";
    string public tokenAbbrv ="MTA";
    uint public totalSupply =0;
    // mapping variable here
    mapping (address => uint) public balances;
    // mint function
    function mint(address _address, uint _value) public {
        totalSupply += _value;
        balances[_address] += _value;
    }
    // burn function
    function burn (address _address,uint _value) public {
        if(balances[_address]>= _value){
        totalSupply -= _value;
        balances[_address] -= _value;
        }
    }
}